Yet more work on the self-consistent correlation potential solution with a wma space.
Changed the motif of the localintegrals.py modifications from "project the operator into the working space"
to "represent the operator in the working space" because I was afraid that the wrong orbitals would end up occupied
and I have no way to constrain that over the course of an scf. Also projected the error-function derivative
so that the derivative has zero value in the non-working space.