Remove testing calculations; improve comments
I can't tell whether numpy is actually multithreading
in tensordot (in otpd.py and pdft_veff.py). When I save an array
to disk and load it and do the tensordot, it appears to
multithread, but in the application it appears not to based on
time reports. This applies even to 10-20 second operations so it
doesn't look like a clock bug.