First successful test-LASSCF calc'n me2n2 (CASSCF)
It was necessary to divide kappa by 2 in the mo update line
in order to make the SCF cycle converge. I need to understand this.
I also need to work out whether Qiming is right about HCI diag
after all. Other than this, all sectors of the Hessian are
accounted for.