Better debug options for lasscf_o1
debug_lasscf_o1 and debug_sa_lasscf_o1 both pass with _debug_o0,
which means the limit of no approximation is both correct for a
single state and averages correctly over multiple states.
debug_sa_lasscf_o1 also passes with _debug_full_pspace, and
debug_lasscf_o1 fails. This is expected but the sector (h.x_ua)_cv
specifically is incorrect, which was unexpected and may be a
problem.