Example debug code + logger convention + bugfix
I added an 'if __name__ == "__main__":' block to the end of
cmspdft3.py, which is code that is only executed when the file
itself is run like "python cmspdft3.py". This code shows how to
look at the NOs of the intermediate states. I also changed the
"print" commands in cmspdft3.kernel to PySCF logger commands and
tried to explain in comments how to use them. I also edited the
code itself by removing the running summation of t, which I think
isn't correct, and set it up to return without crashing even if
it doesn't converge after maxiter (it now returns a boolean that
tells you whether or not it worked).