State rotation Hessian error fix?
I think both the transition density matrices and the "ordinary"
density matrices need to be from the ~rotated~ CI vectors in the
w_klmn function, as when you compute the gradient. Previously,
the w_klmn function computed casdm1 but then didn't turn it into
the AO basis and just passed the previously-computed dm1 arrays
based on the original, unrotated CI vectors.