SA-mix issues in MCPDFT.py
It is impossible to figure out what neleca, nelecb is just from
ncas and the size of the CI array! I rewrote some stuff so that
the user has to submit the full list of ci vectors and a "state"
index for correct handling of SA stuff. It's annoying but it's the
only way to go to SA mix properly. Now I need to fix the
decomposition