MS-PDFT (sipdft) grad singular Hessian handle
Remove singularities from model-space frame-rotation Hessian when
solving the response equations for MS-PDFT analytic molecular
gradients. This puts a floor on how tightly we can solve the
response equations which leads to an error bar on the gradient
(but the alternative is nonsense). If there is a finite gradient
along a singular mode, analytic gradients are unobtainable.