Improved OT fnal string parsing
Looks closely at libxc.py to figure out whether a single-component
functional is exchange or correlation, and errors out if it's
kinetic or a LibXC built-in hybrid. NotImplementedError for stuff
like 'tB3LYP'. You can still type in the formula for 'tB3LYP' in
terms of B88, VWN etc. if you know exactly what you mean by that,
but unlike 'tPBE0' it doesn't have an obvious meaning in the
context of translated hybrid functionals.