Implement AMFI density matrix
There is still a disagreement between OpenMolcas and PySCF despite
fixing the AMFI density matrix. In the H2O example; the SOC Hamiltonian
appears to be off by a factor of 2. On the other hand, in the NpO2
example, the SOC Hamiltonian appears to be off by approximately, but
not exactly, sqrt (2). It's still unclear what the actual problem is.