Improve test_lassi_soc.test_soc_1frag
Fix the relative signs of the wave functions before comparing. This
specifically validates elements [5,1] and [5,3] because the signs
of states 1, 3, and 5 are fixed to reproduce the signs of elements
[1,0], [3,0], and [5,0]. The last of these involve the Sz operator,
while all others involve S+ and S-, so this further validates the
the relative signs of the S+, S-, and Sz implementations. I also
aligned the signs of states 4 and 6 relative to state 2, although
this validates nothing because there are only two nonzero elements
involving those three states; it just makes it more consistent and
elegant.