lasscf_async_combine san check: orth.vec_schmidt?
I've been assuming that orth.vec_schmidt works like np.linalg.qr,
in that if the first x columns are orthonormal, then those columns
are unchanged. I hope that's true but if it's not it will cause
mo_coeff to become non-orthonormal and dm_core to become
non-idempotent. Put a warning in lasscf_async_combine to check.