Fix simplified lasscf_async_combine.orth_orb
The mo_coeff array I tried to use as an orthonormal basis in the
last commit wasn't actually orthonormal because I was putting the
updated mo_cas in the ncore:nocc columns, so it wasn't actually
the last-macrocycle mo_coeff array. Fix and remove pointless
copying of last-macrocycle mo_coeff.