lassi excitation solver api tweak + docstring
set_excited_fragment_ now takes the value of the quantum numbers
instead of the delta from the q-space, since there will be multiple
q-spaces in the future. "excited" is now used in place of "active"
throughout the code to clarify the relationship between the CAS and
the space of excited fragments. docstring for set_excited_fragment_