1st successfully J<0 alfefe LASSI (11,10) nmax=4
Using the ExcitationSolver. Things that enabled this are 1) loosen
self-consistency requirement of initial guess H(E)|Psi> = E|Psi>
and 2) do not greedily solve H(E)|Psi> = E|Psi> for each fragment,
but instead only do one pass of the undressed FCISolver kernel per
VRVSolver kernel call. It is still very slow. As I learned in grad
school, the fixed point algorithm for this fundamentally DOES NOT
WORK when coupling is strong, at least as far as the energy
determination is concerned. The good news is that H(E)|Psi> = E|Psi>
is a strictly one-dimensional monotonic function between
singularities so rootfinding it more properly should be easy.