Remove sort_ci0 step from lassis?
This solves issue #55 because the sort_ci0 step uses a different
objective function than the VRVSolver to sort the states. Even
though LASSIS uses a state-averaged approach, the identity of the
ground state matters because the H(E)|0>=E|0> equation can only
be solved for the ground state, so swapping them willy-nilly
prevents the solver from ever converging. However, this solution
also causes one of the unittests in test_excitations to fail.