LASSIS issue #55 bugfix Hamiltonian update
Correctly update the fragment Hamiltonians after sort_ci0 step in
project_hfrag in the excitation solver in LASSIS. Previously, the
trial energy for the H(E)|Psi>=E|Psi> self-consistent equation was
not updated, leading to meaningless gradient vectors in the
convergence checker whenever the order of states was permuted,
resulting in false convergence failures. The sort_ci0 step is still
highly dangerous because it chooses the state order based on a
different function than the VRVSolver optimizes, but at least the
issue #55 example seems to converge now.