first lasscf_async combine_o1 3-frag convergence
The problem in the previous commit was actually due to the
evaluation of the convergence tolerance. In the subproblem, the
gradient was lower than the tolerance, but the overall gradient
norm was larger than it, so the iteration just stopped making
progress. Enforce a minimum of 1 cycle through the subproblem
iteration brute-force solves this issue for now.