lasscf_async combine_pair_max_frags member option
The algorithm converges massively faster if you optimize all
active-active relaxations simultaneously. However, this is maybe
not very-long-term-scalable because it involves Nroots*Nact^2 and
Nact^4 arrays. So the compromise is we can limit it to a certain
number of fragments at a time. It may be better to choose them
based on the Hessian than the gradient.