lassipdft simplify calling and memory management
Simplify the calling signature of mcpdft.LASSI so that you don't
have to type in ncas_sub and nelecas_sub again. Also implement a
memory management block loop in the lassipdft RDM constructor
function; if the memory is enough it should still do all the states
at once, but if not it'll split the states up by blocks.