LASSIS Sz in ch fbf (#125)
Attempt to solve #125 by imposing some logic on the which Hilbert
spaces will be used. A "target" Sz space is defined by taking the
original reference state and either incrementing or decrementing
the donor and acceptor fragments Sz value based on whether S2
increases or decreases for each fragment and on whether the
interaction in the reference state is ferromagnetic,
antiferromagnetic, or nonmagnetic. Then the fragments are sorted
with the environmental fragments first in increasing order of S2,
then the donor and acceptor fragments in increasing order of S2.
For each candidate Sz space, the difference from the target Sz
vector is evaluated in the fragment order thus described, and the
candidate space which is closest overall to the target space is
chosen.