lassi hci harsh discrimination
I had overlooked how the get_fdm function sums over "equivalent"
excitations and thereby mixes up columns of the final hci_fr table.
An earlier commit fixed a similar problem for the constant terms,
but I just realized that this also applies to the operator terms.
The simplest way around this is to force all bra rootspaces to be
considered as distinct, which as a silver lining allows me to
remove some of the "sum_bra" spaghetti. Fortunately, I only need to
do this for the purposes of testing.