lass lsf pspace ham refactor debug
The culprit was "braket_tab", aka the values of self.nonuniq_exc.
When using spin-orbit coupling, these include pairs of rootspaces
in which m for spectator fragments differ. In the HamS2Ovlp class
and matrix-vector calculator, these terms are multiplied by an
overlap matrix that evaluates to zero. However, in my pspace code,
I had aggressively optimized under the assumption that the
spectator fragment m vectors (and therefore also the overlap matrices)
were idential for all rows of braket_tab. This led to double-
counting in the spin part of fdm, and missing contributions in the
spatial part when a vanishing overlap was taken as the prototype
for a whole block of nonvanishing overlap matrices.