debug lassi test_soc.py hdiag in orthbasis failure
The assumption in lassi.op_o1.hsi.HamS2OvlpOperators.
hdiag_orth_split_braket_tabs, that the coverage by braket_tab of
m cases was always the same for different spincase_keys, sometimes
failed, because the assessment of which fragment statelets were
equivalent to which other fragment statelets under spin flip is
carried out in the OrthBasis object but not in the FragTDMInt
object under SOC conditions. This led the HamS2OvlpOperators
object to separate certain OpTerms which OrthBasis saw as the same
term, leading to double counting. Also, equivalence is evaluated in
two separate places under two separate conditions which don't always
agree, so the result wasn't stable.

I don't know yet if this has anything to do with the tetrahedron
failures.